+++ CHANGELOG.md
@@ -8,6 +8,9 @@ All notable changes to this project are documented in this file.
 
 - basic webpack and tsconfig
 - basic index.html and App
+- ErrorBoundary view and 404 page
+- custom Button component
+- custom Icons for loading, error, and success states
 
 ### Changed